Allow for events when gui is closed
You can now specify what should happen when a player closes a gui. You
can do this either by setting it directly by calling setOnClose on the
gui object or you can set an onClose attribute to the gui class which
links to a method. The method can either have zero parameters, or one
parameter of which the parameter is an InventoryCloseEvent.