Rename attributes to properties
An attribute is key terminology used in XML files which describes an
additional part of information for a tag. Our attributes meant something
different and this could lead to confusion by some users. In order to
avoid this attributes have been renamed to properties, since a property
isn't key terminology used in XML files.