Add skull meta
Forgot to keep this thing compatible with Building Game features. You can now specify skull as additional item meta with owner or id as attribute. Owner will set the specified value as the owner of the skull and id will load the specified id from Mojang's texture servers and will set the skull's texture to that one. Can only be done on a skull item.