Allow panes to overflow
Panes can now be placed/overflow into the bottom half of the inventory,
essentially adding four rows to the previously maximum of six. If any
pane is in the bottom half of the inventory, the player's inventory will
be cached and cleared and will be restored if either the gui is closed
in some way or no panes are in the bottom pane after updating the gui.