Rework item clicking
Items now have a UUID stored so that we can map an ItemStack to a
GuiItem, without needing to know the position or properties of the
ItemStack and GuiItem.

Main reason for changing this is that now events assigned to an item via
GuiItem are only coupled by the item, whereas previously they were
coupled by both the item and the slot, so when you moved an item, the
event would no longer fire when clicking on that item.