Handle items being picked up (#61)
When an item is now being picked up by a human entity who has an open gui which spans across the bottom inventory, items will now be moved to the internal cache if possible instead of picking put into the bottom inventory. Items that can't be put into the cache are left on the ground.